feat(test-infra): enable gRPC by default in TestCluster and share the gRPC builder helpers
Phase 2 of the gRPC transaction-building migration (#12096), closes #12098.

- TestClusterBuilder now enables the fullnode gRPC API by default
  (opt-out via with_fullnode_enable_grpc_api(false)), so grpc_url() is
  always live and WalletContext::get_grpc_client() works out of the box
- add TestCluster::grpc_client() and TestCluster::grpc_transaction_builder()
- promote the per-test move_call_tx / split_coin_equal_tx helpers from
  #11819 into iota-test-transaction-builder, with gas-coin selection
  centralized in select_gas_coin()
- migrate coin_index_test, indexer_api, and iota-source-validation tests
  to the shared helpers; drop the grpc_url threading in source-validation
- remove now-redundant with_fullnode_enable_grpc_api(true) call sites
