+++ crates/iota-cluster-test/src/cluster.rs
@@ -183,8 +183,7 @@ impl Cluster for LocalNewCluster {
 
         let mut cluster_builder = TestClusterBuilder::new()
             .enable_fullnode_events()
-            .with_data_ingestion_dir(data_ingestion_path.clone())
-            .with_fullnode_enable_grpc_api(true);
+            .with_data_ingestion_dir(data_ingestion_path.clone());
 
         // Check if we already have a config directory that is passed
         if let Some(config_dir) = options.config_dir.clone() {
