+++ crates/iota-graphql-rpc/src/test_infra/cluster.rs
@@ -256,8 +256,7 @@ async fn start_validator_with_fullnode(
                 gas_amounts: vec![DEFAULT_GAS_AMOUNT; GAS_OBJECT_COUNT],
             };
             ACCOUNT_NUM
-        ])
-        .with_fullnode_enable_grpc_api(true);
+        ]);
 
     if let Some(internal_data_source_rpc_port) = internal_data_source_rpc_port {
         test_cluster_builder =
