+++ crates/iota-indexer/tests/common/mod.rs
@@ -148,9 +148,7 @@ pub async fn start_test_cluster_with_read_write_indexer(
     pruning_options: Option<PruningOptions>,
 ) -> (TestCluster, PgIndexerStore, HttpClient) {
     let database_name = database_name.into();
-    let mut builder = TestClusterBuilder::new()
-        .with_fullnode_enable_grpc_api(true)
-        .disable_fullnode_pruning();
+    let mut builder = TestClusterBuilder::new().disable_fullnode_pruning();
 
     if let Some(builder_modifier) = builder_modifier {
         builder = builder_modifier(builder);
