+++ crates/test-cluster/Cargo.toml
@@ -24,12 +24,14 @@ iota-config.workspace = true
 iota-core.workspace = true
 iota-framework.workspace = true
 iota-genesis-builder.workspace = true
+iota-grpc-client.workspace = true
 iota-json-rpc-api.workspace = true
 iota-json-rpc-types.workspace = true
 iota-keys.workspace = true
 iota-node.workspace = true
 iota-protocol-config.workspace = true
 iota-sdk.workspace = true
+iota-sdk-transaction-builder.workspace = true
 iota-sdk-types.workspace = true
 iota-swarm.workspace = true
 iota-swarm-config.workspace = true
